backend/requirements.txt
fastapi
uvicorn
dspy-ai
pydantic
python-dotenv
httpx
google-generativeai
openpyxl
